updated by rafael

+++ Meta 2/TP/barraAco.cpp
@@ -23,7 +23,7 @@ double BarraAco::producaoBAF() {
     return custoFerro;
 }
 
-void BarraAco::aumenta(int quant) {
+void BarraAco::aumenta(double quant) {
     quantidade += quant;
 }
